requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
Faker==9.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3

